Fix multiple OR facets with database backend
This fixes an issue where the result count for facets would be incorrect
when using multiple OR facets. It is due to shallow-cloning the search
query, looping through all the facets and unsetting conditions for each
facet. The conditions are only supposed to be removed for a specific
facet for a single interation of the loop, but due to the shallow clone,
the conditions remain removed for the successive iterations. This causes
later facets to lose the previous facets conditions, causing the result
count for those facets to be incorrect.